Moved tests to new file micer_tests.py in tests folder and fixed test_add_micer_settings
